% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\docType{package}
\name{lumberjack}
\alias{lumberjack}
\alias{lumberjack-package}
\title{The pipe operator that logs}
\description{
The pipe operator that logs
}
\section{Overview}{


The lumberjack \code{\%L>\%} behaves much like other function
composition ('pipe') operators available in R (e.g. \href{https://CRAN.R-project.org/package=magrittr}{magrittr}
, \href{https://github.com/piccolbo/yapo}{yapo}, \href{https://CRAN.R-project.org/package=pipeR}{pipeR})
with one exception: it allows for logging the changes made to the data
by the functions acting on it.

The actual logging mechanism is completely flexible and extensible. This
package comes with a few predefined loggers, but users and package authors
can write their own logger that follows the lumberjack API.

See the \href{../doc/intro.html}{Introductory vignette} to start logging
or the \href{../doc/extending.html}{extending lumberjack} manual to start 
writing your own loggers.

Happy logging!
}

