% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\name{start_log}
\alias{start_log}
\title{Start logging changes on a dataset.}
\usage{
start_log(data, log = simple$new())
}
\arguments{
\item{data}{An R object carrying data.}

\item{log}{A logging object (typically an environment wrapped in an S3 class)}
}
\description{
Start logging changes on a dataset.
}
