% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\name{stop_log}
\alias{stop_log}
\title{Stop logging}
\usage{
stop_log(data, ...)
}
\arguments{
\item{data}{An R object carrying data}

\item{...}{Passed to the logger's \code{stop} method, if it exists.}
}
\description{
Calls the logger's \code{$stop()} method if it exists, and removes
the logger as attribute from \code{data}.
}
