% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\name{dump_log}
\alias{dump_log}
\title{Dump a log}
\usage{
dump_log(data, stop = TRUE, ...)
}
\arguments{
\item{data}{data set containing a log}

\item{stop}{stop logging after the dump?}

\item{...}{Arguments past to the \code{dump} method of the logger.}
}
\value{
The data, invisibly
}
\description{
Dump a log
}
