% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\name{get_log}
\alias{get_log}
\title{Get log object from a data item}
\usage{
get_log(data)
}
\arguments{
\item{data}{An R object carrying data.}
}
\value{
A logging object, or \code{NULL} if none exists.
}
\description{
Get log object from a data item
}
