% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\docType{package}
\name{lumberjack}
\alias{lumberjack}
\alias{lumberjack-package}
\title{Track changes in data}
\description{
This package allows you to track changes in R objects by defining one or
more loggers for each object. There are a number of built-in loggers and
users (or package authors) can create their own loggers.  To get started
please have a look at the \href{../doc/using_lumbjerjack.pdf}{using
lumberjack} vignette.
}
\author{
Mark van der Loo
}
